% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoB_detection.R
\name{Extract_RoB}
\alias{Extract_RoB}
\title{RiskOfBiasIdentification}
\usage{
Extract_RoB(
  searchingData = NULL,
  dictionary = "data/Preclinical_RoB_regex.csv",
  linkSearchHeaders = "",
  dictionaryNameHeader = "Name",
  dictionaryRegexHeader = "Regex",
  ignoreCase = TRUE,
  ignoreExistingTextFile = TRUE,
  cutIntro = FALSE,
  cutRefs = FALSE,
  extractStrings = FALSE,
  conversionSoftware = "pdftotext"
)
}
\arguments{
\item{searchingData}{Either a dataset or a link to the dataset to search from}

\item{dictionary}{Either a dictionary dataset, or a link to the dictionary dataset to run the function on.
It should consist two columns: name of the term and search string of the term. Regular expression (Perl) is accepted for the search string.
If there is only one column, that column will be used both as name and regular expression.}

\item{linkSearchHeaders}{A list of the headers of the columns to read and search from. A list of character. Default value is empty string}

\item{dictionaryNameHeader}{The header string of name column in dictionary}

\item{dictionaryRegexHeader}{The header string of regular expression column in dictionary}

\item{ignoreCase}{boolean to decide whether to ignore the case in searching the content in dictionary in the searchingData or not}

\item{ignoreExistingTextFile}{ignoreExistingTextFile}

\item{cutIntro}{boolean varible to indicate whether introduction/background should be removed from pdfs}

\item{cutRefs}{boolean varible to indicate whether reference section should be removed from pdfs}

\item{extractStrings}{= boolean variable to indicate whether matching strings from full text should be extracted}

\item{conversionSoftware}{Software used to covert pdf to text. Default value is 'pdftotext'}
}
\value{
result flags
}
\description{
Identify Risk of Bias in input data within columns 'Title', 'Abstract' and 'PdfRelevantPath'
}
