% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoB_detection.R
\name{ValidateDictionary}
\alias{ValidateDictionary}
\title{ValidateDictionary}
\usage{
ValidateDictionary(myDictionary, dictionaryNameHeader, dictionaryRegexHeader)
}
\arguments{
\item{myDictionary}{my Dictionary}

\item{dictionaryNameHeader}{dictionary Name Header}

\item{dictionaryRegexHeader}{dictionary Regex Header}
}
\value{
valid myDictionary
}
\description{
Validate Dictionary
}
